cosmos: relocate live 1008/3 tests to multi_region/multi_write modules
The two live-account regression tests for 403/1008 and 403/3 were
sitting in emulator_tests/driver_fault_injection.rs even though they
require a real multi-region (1008) or multi-write (403/3) Cosmos
account — the local emulator cannot satisfy that, so they could only
ever fail in Build Test stages that ran the emulator suite on Windows.

Move them to the modules whose conditional gating already targets the
right cosmos-weekly pipelines:

  multi_region_tests/live_403_1008_topology_refresh.rs   (1008)
  multi_write_tests/live_403_3_failover.rs               (403/3)

Each file is feature-gated on fault_injection and per-test gated on
the matching test_category so the right cosmos-weekly stage picks
them up.
