+++ sdk/cosmos/azure_data_cosmos_driver/tests/multi_region_tests/mod.rs
@@ -5,3 +5,4 @@
 
 mod driver_partition_failover;
 mod hedging;
+mod live_403_1008_topology_refresh;
