+++ sdk/cosmos/azure_data_cosmos_driver/tests/multi_write_tests/mod.rs
@@ -2,3 +2,4 @@
 // Licensed under the MIT License.
 
 pub mod driver_partition_failover;
+pub mod live_403_3_failover;
